Implement simplistic word list persistence service.
Once this is integrated with the controller, the user will be able to
add new words to the list and see them reflected in the app.

In a real web app, this would be backed by a database; our very simple
implementation is instead backed by a list.

Also liberalize style rules to allow for adjacent declarations without
intervening blank lines (for use especially in interface declarations).