Add missing `@ResponseStatus` annotation for Ajax call handler
The call to post a word is a REST handler that should be returning a
generic "OK" status, but instead, it seems to be assuming that the URL
should be treated as a view name (which then fails to resolve, resulting
in a 500 Internal Server Error status).  This doesn't seem to affect the
functionality, but is bad practice.

To build the regression test for this, we need a way to distinguish the
bad call (which nevertheless seems to have an "OK" status, even though
the subsequent attempt to render the view will fail).  Checking that
there is no model or view attached seems to be the only way to do this,
and there's no existing matcher for that.  We could handle this with a
lambda, but we choose to instead follow Spring's existing patterns
(drill down to the model-and-view matchers via a static method, then
build the actual assertion onto the resulting object).